Move massUpdate into normal soupCal service, rebolster the testing for cronHelper
