+++ test/api/helpers/cronHelperTest.js
@@ -9,7 +9,10 @@ const testHelper = require('../../helper/testHelper');
 const cronHelper = require('../../../api/helpers/cronHelper');
 const parseSubito = require('../../../api/helpers/parseSubito');
 const slack = require('../../../api/helpers/slack');
+const logger = require('../../../api/helpers/logger');
+const soupCalendarViewService = require('../../../api/services/soupCalendarViewService');
 
+const testSubscribers = require('../../data/Subscribers.json');
 const testHtml = fs.readFileSync(path.join(__dirname, '../../data') + '/test-calendar.html', 'utf-8');
 
 describe('cronHelper', () => {
@@ -17,13 +20,17 @@ describe('cronHelper', () => {
   after(testHelper.clearData);
   describe('importCalendar', () => {
     it('should not error', (done) => {
+      const loggerSpy = sinon.spy(logger, 'info');
       sinon.stub(parseSubito.private, 'fetchSoupPage').yields(null, null);
       sinon.stub(fs, 'readFile').yields(null, testHtml);
       cronHelper.importCalendar(testHelper.db)((err, result) => {
         should.not.exist(err);
         assert(result.rows > 0);
+        assert(loggerSpy.calledWith('importCalendar complete:: '));
+        
         parseSubito.private.fetchSoupPage.restore();
         fs.readFile.restore();
+        loggerSpy.restore();
         done();
       });
     });
@@ -31,10 +38,17 @@ describe('cronHelper', () => {
 
   describe('processSubscribers', () => {
     it('should not error', (done) => {
-      sinon.stub(slack, 'messageUser').yields(null, { status: 'Success' });
+      const slackSpy = sinon.stub(slack, 'messageUser').yields(null, { status: 'Success' });
+      const loggerSpy = sinon.spy(logger, 'info');
+      const soupCalSpy = sinon.spy(soupCalendarViewService, 'getSoupsForDay');
       cronHelper.processSubscribers(testHelper.db)((err) => {
         should.not.exist(err);
-        slack.messageUser.restore();
+        assert(soupCalSpy.calledOnce);
+        assert(loggerSpy.calledWith('processSubscribers complete'));
+        assert.equal(slackSpy.getCalls().length, testSubscribers.length);
+
+        slackSpy.restore();
+        soupCalSpy.restore();
         done();
       });
     });
