+++ test/api/services/soupCalendarServiceTest.js
@@ -1,7 +1,9 @@
 'use strict';
 
 const assert = require('assert');
+const should = require('should');
 const moment = require('moment');
+const lodash = require('lodash');
 
 const testHelper = require('../../helper/testHelper');
 const soupCalendarService = require('../../../api/services/soupCalendarService');
@@ -95,6 +97,40 @@ describe('soupCalendarService', () => {
       soupCalendarService.searchForSoup(testHelper.db, searchStr, (err, res) => {
         assert.equal(err, null);
         assert.equal(res.length, 0, 'should not have results');
+        done();
+      });
+    });
+  });
+
+  describe('massUpdate', () => {
+    it('should get soup calendar entry for day', (done) => {
+      const updates = [];
+      const soupOptions = ['Chicken Noodle', 'Beef Stew',
+        'Turkey Bean Soup', 'Black Bean (gf)', 'Italian Wedding (gf)',
+        'Local Corn Chowder'];
+      let generateDays = lodash.random(1, 20);
+      let expectedStartDate;
+      let expectedEndDate;
+      for (var i = 0; i < generateDays; i++) {
+        const soups = lodash.clone(soupOptions).splice(lodash.random(0, soupOptions.length - 2), 2);
+        const date = moment().add(lodash.random(-10, 10), 'd');
+
+        if (!expectedStartDate) expectedStartDate = date;
+        else if (date < expectedStartDate) expectedStartDate = date;
+        if (!expectedEndDate) expectedEndDate = date;
+        else if (date > expectedEndDate) expectedEndDate = date;
+
+        updates.push({
+          date: date.format(),
+          soups: soups
+        });
+      }
+
+      soupCalendarService.massUpdate(testHelper.db, updates, (err, updated) => {
+        should.not.exist(err);
+        updated.should.have.property('rows', updates.length * 2);
+        updated.should.have.property('startDate', expectedStartDate.format('YYYY/MM/DD'));
+        updated.should.have.property('endDate', expectedEndDate.format('YYYY/MM/DD'));
         done();
       });
     });
